fix(data): restore disappear key round corners (#895)
v3.2.10 heavily refactors the parsing of the theme, replacing the Drawable type of plain-color-only graphics / rects with ColorDrawable instead of GradientDrawable, in order to reduce the performance loss.
However, that is just my wishful thinking, as there is no conclusive evidence show that that way can do so here.
ColorDrawable is a simple Drawable type, which is basically just for showing a plain color shape, but without round corner setting. So let us go back to GradientDrawable.