refactor(data,ui): hardcode the shared data dir to the external files dir
According to the librime's design, the shared data dir "holds files that are shared by multiple users on the machine, usually written by an input method setup program. The Rime input method, when searching for a resource, gives priority to the files in the user's folder. When the user file does not exist, look for it in the shared data dir ..." [1], "... contains the source files of the builtin input schemas. These files are a part of Rime or its distributions and are read-only to users on systems with tight access controls, so any changes other than software version updates are rejected. Once a user modifies a file here, it's likely to affect subsequent software upgrades or lose data after the upgrade."[2]

Ref: 1. https://github.com/rime/home/wiki/SharedData
2. https://github.com/rime/home/wiki/RimeWithSchemata#rime-%E4%B8%AD%E7%9A%84%E6%95%B8%E6%93%9A%E6%96%87%E4%BB%B6%E5%88%86%E4%BD%88%E5%8F%8A%E4%BD%9C%E7%94%A8