Security risk: Disable security and privacy disrespecting `draft`, `collection`, `clipboard` keyboard in `liquid_keyboard` by default
When using trime as keyboard to unlock your phone in lockscreen, the password is recorded in `draft` keyboard. If you enable `draft` keyboard and your phone is stolen, other people can simply unlock your phone by using your recent input in the `draft` keyboard. For security reason, this should be disable by default.

For the privacy reason, `clipboard` and `collection` should also be disabled because it can show sensitive information in the lockscreen,