refactor: display `InitialKeyboard` before deployment or lacks of permissions
If permissions are not granted, `RimeWrapper.canStart` will be set to false, so
`RimeWrapper` will save the `Runnable` and wait to start later.  When permission are
granted, `RimeWrapper.canStart` will be set to true in `onCreate()`, `onCreateInputView()`
and `onStartInputView()`.  And rime deployment will be triggered accordingly.

For UI display:  If permissions are not granted, an error message will be displayed
in `InitialKeyboard`.  If permissions are granted, `InitialKeyboard` with deployment
progress will be displayed.  If deployment is completed, the normal keyboard will be
displayed.

Do not set `initialKeyboard` to null to prevent NPE. As the callback in `RimeWrapper`
maybe perform faster or slower, we will not be sure when `setInputView()` will run.
We should either check for null for all call to `initialKeyboard`, or do not set it
to null.

Refs #1159