refactor: use Jackson to deal with flexible theme parsing
KotlinX Serialization (kaml) is nice, but it's too oriented to parse our themes. Although we have wrapped it with custom Config class, it slows down the performance. Jackson is a popular library to parse JSON, YAML and more, and it provides quite performance as KotlinX Serialization, and it can be very flexible to deal with complex data parsing.